third_party/WebKit/LayoutTests/http/tests/devtools/console/console-retain-autocomplete-on-typing-expected.txt

Verify that console does not hide autocomplete during typing.


Running: testSummonSuggestBox
Suggestions shown.

Running: testTypeText
SUCCESS: suggestbox is not hidden during typing.
